Finding cell type specific differentially methylated regions

This function is modified from methylCC to use flow sorted array reference methylomes for either blood, cord blood, or brain to identify cell type specific differentially methylated regions.

find_dmrs2(
  verbose = TRUE,
  gr_target = NULL,
  include_cpgs = FALSE,
  include_dmrs = TRUE,
  num_cpgs = 50,
  num_regions = 50,
  bumphunter_beta_cutoff = 0.2,
  dmr_up_cutoff = 0.5,
  dmr_down_cutoff = 0.4,
  dmr_pval_cutoff = 1e-11,
  cpg_pval_cutoff = 1e-08,
  cpg_up_dm_cutoff = 0,
  cpg_down_dm_cutoff = 0,
  pairwise_comparison = FALSE,
  mset_train_flow_sort = NULL
)

Arguments

verbose
TRUE/FALSE argument specifying if verbose messages should be returned or not. Default is TRUE.

gr_target
Default is NULL. However, the user can provide a GRanges object from the object in estimatecc. Before starting the procedure to find differentially methylated regions, the intersection of the gr_target and GRanges object from the reference methylomes (FlowSorted.Blood.450k).

include_cpgs
TRUE/FALSE. Should individual CpGs be returned. Default is FALSE.

include_dmrs
TRUE/FALSE. Should differentially methylated regions be returned. Default is TRUE. User can turn this to FALSE and search for only CpGs.

num_cpgs
The max number of CpGs to return for each cell type. Default is 50.

num_regions
The max number of DMRs to return for each cell type. Default is 50.

bumphunter_beta_cutoff
The cutoff threshold in bumphunter() in the bumphunter package.

dmr_up_cutoff
A cutoff threshold for identifying DMRs that are methylated in one cell type, but not in the other cell types.

dmr_down_cutoff
A cutoff threshold for identifying DMRs that are not methylated in one cell type, but methylated in the other cell types.

dmr_pval_cutoff
A cutoff threshold for the p-values when identifying DMRs that are methylated in one cell type, but not in the other cell types (or vice versa).

cpg_pval_cutoff
A cutoff threshold for the p-values when identifying differentially methylated CpGs that are methylated in one cell type, but not in the other cell types (or vice versa).

cpg_up_dm_cutoff
A cutoff threshold for identifying differentially methylated CpGs that are methylated in one cell type, but not in the other cell types.

cpg_down_dm_cutoff
A cutoff threshold for identifying differentially methylated CpGs that are not methylated in one cell type, but are methylated in the other cell types.

pairwise_comparison
TRUE/FAlSE of whether all pairwise comparisons (e.g. methylated in Granulocytes and Monocytes, but not methylated in other cell types). Default if FALSE.

mset_train_flow_sort
Default is NULL and will select select "FlowSorted.Blood.450k". Alternatively, the user can specify another existing dataset c("FlowSorted.Blood.EPIC", "FlowSorted.CordTissueAndBlood.EPIC", "FlowSorted.CordBlood.450k", "FlowSorted.DLPFC.450k"). Additionally, a user can provide a MethylSet object after processing the FlowSorted.Blood.450k dataset. The default normalization is preprocessQuantile().

Value

A list of data frames and GRanges objects.