Preprocess cytosine reports

Process bismark cytosine reports into bsseq objects with design matrix pData

processBismark(
  files = list.files(path = getwd(), pattern = "*.txt.gz"),
  meta = openxlsx::read.xlsx("sample_info.xlsx", colNames = TRUE) %>%
    dplyr::mutate_if(is.character, as.factor),
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL,
  coverage = coverage,
  cores = cores,
  perGroup = perGroup,
  sexCheck = FALSE
)

Arguments

files
List of cytosine report file paths

meta
Design matrix data frame with sample name in the Name column

testCovariate
Factor of interest

adjustCovariate
Variables to adjust for

matchCovariate
Variable to block for when constructing permutations

coverage
CpG coverage cutoff (1x recommended)

cores
Integer specifying the number of cores to use

perGroup
Percent of samples per a group to apply the CpG coverage cutoff to (from 0 to 1)

sexCheck
Logical (TRUE or FALSE) indicating whether to confirm the sex of samples and drop the sex chromosomes if both sexes are present. This function requires a column called "Sex" (case sensitive) in sample_info.xlsx. Males should be coded as either "Male", "male", "M", or "m". Females coded as "Female", "female", "F", or "f".

Details

processBismark